+++ xbmc/pvr/guilib/guiinfo/PVRGUIInfo.cpp
@@ -1482,7 +1482,7 @@ bool CPVRGUIInfo::GetPVRInt(const CFileItem* item, const CGUIInfo& info, int& iV
         iValue = 0xFF;
       return true;
     case PVR_CLIENT_COUNT:
-      iValue = CServiceBroker::GetPVRManager().Clients()->EnabledClientAmount();
+      iValue = static_cast<int>(CServiceBroker::GetPVRManager().Clients()->EnabledClientAmount());
       return true;
   }
   return false;
